+++ src/Platforms/CUDA/CMakeLists.txt
@@ -17,6 +17,5 @@ if(NOT QMC_CUDA2HIP)
   target_link_libraries(platform_cuda PRIVATE ${CUDA_LIBRARIES} ${CUDA_cusolver_LIBRARY})
 else()
   hip_add_library(platform_cuda ${CUDA_SRCS})
-  target_link_libraries(platform_cuda PUBLIC hip::host roc::hipblas roc::rocsolver)
-  target_include_directories(platform_cuda PUBLIC ${PROJECT_SOURCE_DIR}/external_codes/thrust)
+  target_link_libraries(platform_cuda PUBLIC platform_rocm)
 endif()
