+++ src/Platforms/ROCm/CMakeLists.txt
@@ -2,20 +2,16 @@
 #// This file is distributed under the University of Illinois/NCSA Open Source License.
 #// See LICENSE file in top directory for details.
 #//
-#// Copyright (c) 2020 QMCPACK developers.
+#// Copyright (c) 2021 QMCPACK developers.
 #//
 #// File developed by: Ye Luo, yeluo@anl.gov, Argonne National Laboratory
 #//
 #// File created by: Ye Luo, yeluo@anl.gov, Argonne National Laboratory
 #//////////////////////////////////////////////////////////////////////////////////////
 
-set(CUDA_LEGACY_SRCS cuda_inverse.cu gpu_vector.cpp gpu_misc.cpp)
+set(ROCM_SRCS hipBLAS.hip.cpp)
 
-if(NOT QMC_CUDA2HIP)
-  cuda_add_library(platform_cuda_legacy ${CUDA_LEGACY_SRCS})
-  cuda_add_cublas_to_target(platform_cuda_legacy)
-  target_link_libraries(platform_cuda_legacy PRIVATE ${CUDA_LIBRARIES})
-else()
-  hip_add_library(platform_cuda_legacy ${CUDA_LEGACY_SRCS})
-  target_link_libraries(platform_cuda_legacy PUBLIC platform_rocm)
-endif(NOT QMC_CUDA2HIP)
+set_source_files_properties(${ROCM_SRCS} PROPERTIES HIP_SOURCE_PROPERTY_FORMAT 1)
+hip_add_library(platform_rocm ${ROCM_SRCS})
+target_link_libraries(platform_rocm PUBLIC hip::host roc::hipblas roc::rocsolver)
+target_include_directories(platform_rocm PUBLIC ${PROJECT_SOURCE_DIR}/external_codes/thrust)
